requirements.txt
imbalanced_learn==0.12.3
lightgbm==4.6.0
linearmodels==6.1
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
plotly==5.24.1
python_dateutil==2.9.0.post0
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.47.1
statsmodels==0.14.2
streamlit==1.37.1
streamlit_option_menu==0.4.0
